LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpTransceiver.https-expected.txt


Harness Error (FAIL), message = Error: assert_true: ICE failed expected true got false

PASS checkAddTransceiverNoTrack
PASS checkAddTransceiverWithTrack
PASS checkAddTransceiverWithAddTrack
PASS checkAddTransceiverWithDirection
FAIL checkMsidNoTrackId promise_test: Unhandled rejection with value: object "SyntaxError: Expects 2 fields."
PASS checkAddTransceiverWithSetRemoteOfferSending
PASS checkAddTransceiverWithSetRemoteOfferNoSend
PASS checkAddTransceiverBadKind
PASS checkNoMidOffer
PASS checkNoMidAnswer
PASS checkSetDirection
FAIL checkCurrentDirection assert_equals: expected "[{currentDirection:null}]" but got "[{currentDirection:\"inactive\"}]"
TIMEOUT checkSendrecvWithNoSendTrack Test timed out
NOTRUN checkSendrecvWithTracklessStream
NOTRUN checkAddTransceiverNoTrackDoesntPair
NOTRUN checkAddTransceiverWithTrackDoesntPair
NOTRUN checkAddTransceiverThenReplaceTrackDoesntPair
NOTRUN checkAddTransceiverThenAddTrackPairs
NOTRUN checkAddTrackPairs
NOTRUN checkReplaceTrackNullDoesntPreventPairing
NOTRUN checkRemoveAndReadd
NOTRUN checkAddTrackExistingTransceiverThenRemove
NOTRUN checkRemoveTrackNegotiation
NOTRUN checkMute
NOTRUN checkStop
NOTRUN checkStopAfterCreateOffer
NOTRUN checkStopAfterSetLocalOffer
NOTRUN checkStopAfterSetRemoteOffer
NOTRUN checkStopAfterCreateAnswer
NOTRUN checkStopAfterSetLocalAnswer
NOTRUN checkStopAfterClose
NOTRUN checkLocalRollback
NOTRUN checkRollbackAndSetRemoteOfferWithDifferentType
NOTRUN checkRemoteRollback
NOTRUN checkMsectionReuse
NOTRUN checkStopAfterCreateOfferWithReusedMsection
NOTRUN checkAddIceCandidateToStoppedTransceiver
NOTRUN checkBundleTagRejected
